review includes, fix compilation warnings detected by clang
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/triva@1657 96ca98c3-5b99-4e05-a768-75e2d8bf8841

+++ Dot/Dot.m
@@ -71,7 +71,7 @@
   while ((et = [en nextObject]) != nil) {
           if ([self isContainerEntityType:et]) {
       NSColor *color = [et color];
-      float red, green, blue, alpha;
+      double red, green, blue, alpha;
       NS_DURING
       [color getRed: &red green: &green blue: &blue alpha: &alpha];
       NS_HANDLER
